tools/admin-ui/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
httpx>=0.27.0
psycopg2-binary>=2.9.9
